Refactor pitch comparison CLI into modular components
