+++ src/spectrum_analysis/__init__.py
@@ -1,4 +1,5 @@
 """Spectrum analysis utilities and interactive visualizer."""
+
 from .audio_sources import AudioSource, DemoSource, MicSource
 from .cli import build_config, create_source, main, parse_args
 from .compare_pitch_cli import main as compare_pitch_main
