+++ src/spectrum_analysis/utils.py
@@ -1,4 +1,5 @@
 """Utility helpers for spectrogram analysis."""
+
 from __future__ import annotations
 
 import numpy as np
